add basic support for multiple OKAPI installations (see #28)
This is the first step of the refactoring which allows switching between the
German site and a test server during runtime. The configuration values and
methods have been adapted to not rely on a static value anymore, but to call
dedicated methods instead which return the value corresponding to the selected
Opencaching server.

As all the changes are mostly intended for testing the application itself,
especially when it comes to actual write operations (submitting logs), the
new functionality is not availabe from within the GUI. We may want to decide
whether including support for the other (real) OKAPI installations makes sense
and how this can be handled later on.